FA-694-add-roles-to-orchestrator-function-app (#100)
* Add orchestrator access to blob storage with enhanced permissions

This commit introduces a new module for granting the orchestrator access to blob storage, allowing it to write and expose conversation links. The new module, `orchestratorBlobStorageAccess`, is defined in `main.bicep` and utilizes the `blobstorage-access-orc.bicep` file to assign the Storage Blob Data Contributor and Storage Blob Delegator roles to the orchestrator's service principal.

* Update role assignment API version in blobstorage-access.bicep

This commit updates the role assignment resource in the blobstorage-access.bicep file to use the latest API version '2022-04-01'. This change ensures compatibility with the current Azure API standards and may include enhancements or fixes from the previous version.